\name{BayesianMIR-package}
\alias{BayesianMIR-package}
\alias{BayesianMIR}
\docType{package}
\title{
\packageTitle{BayesianMIR}
}
\description{
\packageDescription{BayesianMIR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BayesianMIR}
\packageIndices{BayesianMIR}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{BayesianMIR}

Maintainer: \packageMaintainer{BayesianMIR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }