% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIRfitting.R
\name{getInit}
\alias{getInit}
\title{Generate initial values}
\usage{
getInit(hypara)
}
\arguments{
\item{hypara}{A list created from \code{\link{getHypara}}.}
}
\value{
A list with components:
\item{coef}{Random variables from MVN(\code{hypara$hp_mu_coef}, $10I$).}
\item{sig2_error}{Random variable from Gamma(\code{hypara$hp_a}, \code{hypara$hp_b}).}
\item{delta}{Binary variables. Within a bag $i$, each random vector is generated from Multinomial(1, \code{hp_pi_list[[i]]}).}
}
\description{
Generate initial values for parameters
}
\examples{
getInit(getHypara(Tidy_dataset(label = rnorm(3), feature_inst = replicate(3, matrix(rnorm(10), 2), simplify = FALSE))))
}
