% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BMIRfitting.R
\name{getPar}
\alias{getPar}
\title{Generate all input parameters}
\usage{
getPar(tidydata)
}
\arguments{
\item{tidydata}{A list created from \code{\link{Tidy_dataset}}.}
}
\value{
A list with components:
\item{n}{\code{nsample} in \code{\link{Tidy_dataset}}}
\item{p}{\code{nfeature} in \code{\link{Tidy_dataset}}}
\item{ninst}{\code{ninst} in \code{\link{Tidy_dataset}}}
\item{m}{Total number of instances}
\item{membership}{{\code{membership} in \code{\link{Tidy_dataset}}}}
\item{Y}{{\code{label} in \code{\link{Tidy_dataset}}}}
\item{X1}{Covariate matrix of all instances.}
\item{hp_sig2_y}{Same as in \code{\link{getHypara}}}
\item{hp_mu_coef}{Same as in \code{\link{getHypara}}}
\item{hp_a}{Same as in \code{\link{getHypara}}}
\item{hp_b}{Same as in \code{\link{getHypara}}}
\item{hp_pi}{Same as in \code{\link{getHypara}}}
\item{hp_inv_Sig_coef}{Same as in \code{\link{getHypara}}}
\item{hp_g_coef}{Same as in \code{\link{getHypara}}}
\item{coef}{Same as in \code{\link{getInit}}}
\item{sig2_error}{Same as in \code{\link{getInit}}}
\item{delta}{Same as in \code{\link{getInit}}}
}
\description{
Generate all input parameters required in \code{\link{Run1Gibbs_cpp}}.
}
\examples{
getPar(Tidy_dataset(label = rnorm(3), feature_inst = replicate(3, matrix(rnorm(10), 2), simplify = FALSE)))
}
