% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tidydata.R
\name{Tidy_dataset}
\alias{Tidy_dataset}
\title{Tidy a dataset for the multiple instance learning}
\usage{
Tidy_dataset(label = NULL, feature_inst, membership = NULL)
}
\arguments{
\item{label}{A vector of response variables (of length r). For unlabeled bags, it is set to NULL.}

\item{feature_inst}{Either a list of length r or a matrix. If a list is given, each component contains covariate information of multiple instances in a matrix form (of same column size, but possibly different row size). If a matrix is given, \code{membership} that indicate group indices of all instances should be specified.}

\item{membership}{A vector of integers that indicate group indices of all instances.}
}
\value{
A list with components:
\item{label}{Same as in \code{label}.}
\item{feature_inst}{Same as in \code{feature_inst}.}
\item{nsample}{The number of bags (or samples).}
\item{nfeature_inst}{The dimension of covariates.}
\item{ninst}{A vector of length r, each entry of which is the number of instances in a bag.}
\item{membership}{A vector of integers that indicate group indices of all instances.}
}
\description{
Generate a list of components frquently used in the multiple instance learning
}
\examples{
## List input
### labeled bags
Tidy_dataset(label = rnorm(3), feature_inst = replicate(3, matrix(rnorm(10), 2), simplify = FALSE))
### unlabeled bags
Tidy_dataset(feature_inst = replicate(3, matrix(rnorm(10), 2), simplify = FALSE))
## Matrix input is also possible
feature <- cbind(c(1,1,1,2,2,3), matrix(rnorm(10 * 3), ncol = 5))
Tidy_dataset(label = rnorm(3), feature_inst = feature[,-1], membership = feature[,1])
}
